split the runner, group, and test logic into separate files
