+++ docs/dir_bbea7df79ae1dc059877576c0778bec3.html
@@ -75,7 +75,7 @@ $(function() {
 <tr class="heading"><td colspan="2"><h2 class="groupheader"><a name="files"></a>
 Files</h2></td></tr>
 <tr class="memitem:run_8h"><td class="memItemLeft" align="right" valign="top">file &#160;</td><td class="memItemRight" valign="bottom"><a class="el" href="run_8h.html">run.h</a> <a href="run_8h_source.html">[code]</a></td></tr>
-<tr class="memdesc:run_8h"><td class="mdescLeft">&#160;</td><td class="mdescRight">Defines the runner and group functionality. <br /></td></tr>
+<tr class="memdesc:run_8h"><td class="mdescLeft">&#160;</td><td class="mdescRight">Defines the runner functionality. <br /></td></tr>
 <tr class="separator:"><td class="memSeparator" colspan="2">&#160;</td></tr>
 <tr class="memitem:scrutiny_8h"><td class="memItemLeft" align="right" valign="top">file &#160;</td><td class="memItemRight" valign="bottom"><a class="el" href="scrutiny_8h.html">scrutiny.h</a> <a href="scrutiny_8h_source.html">[code]</a></td></tr>
 <tr class="memdesc:scrutiny_8h"><td class="mdescLeft">&#160;</td><td class="mdescRight">Master include file. <br /></td></tr>
