+++ docs/files.html
@@ -71,9 +71,10 @@ $(function() {
 <div class="levels">[detail level <span onclick="javascript:toggleLevel(1);">1</span><span onclick="javascript:toggleLevel(2);">2</span><span onclick="javascript:toggleLevel(3);">3</span>]</div><table class="directory">
 <tr id="row_0_" class="even"><td class="entry"><span style="width:0px;display:inline-block;">&#160;</span><span id="arr_0_" class="arrow" onclick="toggleFolder('0_')">&#9660;</span><span id="img_0_" class="iconfopen" onclick="toggleFolder('0_')">&#160;</span><a class="el" href="dir_d44c64559bbebec7f509842c48db8b23.html" target="_self">include</a></td><td class="desc"></td></tr>
 <tr id="row_0_0_"><td class="entry"><span style="width:16px;display:inline-block;">&#160;</span><span id="arr_0_0_" class="arrow" onclick="toggleFolder('0_0_')">&#9660;</span><span id="img_0_0_" class="iconfopen" onclick="toggleFolder('0_0_')">&#160;</span><a class="el" href="dir_bbea7df79ae1dc059877576c0778bec3.html" target="_self">scrutiny</a></td><td class="desc"></td></tr>
-<tr id="row_0_0_0_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="run_8h_source.html"><span class="icondoc"></span></a><a class="el" href="run_8h.html" target="_self">run.h</a></td><td class="desc">Defines the runner and group functionality </td></tr>
-<tr id="row_0_0_1_"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="scrutiny_8h_source.html"><span class="icondoc"></span></a><a class="el" href="scrutiny_8h.html" target="_self">scrutiny.h</a></td><td class="desc">Master include file </td></tr>
-<tr id="row_0_0_2_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="test_8h_source.html"><span class="icondoc"></span></a><a class="el" href="test_8h.html" target="_self">test.h</a></td><td class="desc">Defines the test macros </td></tr>
+<tr id="row_0_0_0_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="groups_8h_source.html"><span class="icondoc"></span></a><b>groups.h</b></td><td class="desc"></td></tr>
+<tr id="row_0_0_1_"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="run_8h_source.html"><span class="icondoc"></span></a><a class="el" href="run_8h.html" target="_self">run.h</a></td><td class="desc">Defines the runner functionality </td></tr>
+<tr id="row_0_0_2_" class="even"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="scrutiny_8h_source.html"><span class="icondoc"></span></a><a class="el" href="scrutiny_8h.html" target="_self">scrutiny.h</a></td><td class="desc">Master include file </td></tr>
+<tr id="row_0_0_3_"><td class="entry"><span style="width:48px;display:inline-block;">&#160;</span><a href="test_8h_source.html"><span class="icondoc"></span></a><a class="el" href="test_8h.html" target="_self">test.h</a></td><td class="desc">Defines the test macros </td></tr>
 </table>
 </div><!-- directory -->
 </div><!-- contents -->
