+++ docs/globals.html
@@ -156,23 +156,8 @@ $(function() {
 <li>SCR_RUN_FLAG_FAIL_FAST
 : <a class="el" href="run_8h.html#a4701a8c215ac188f58bead0aaaa2e12d">run.h</a>
 </li>
-<li>SCR_TEST_FLAG_XFAIL
-: <a class="el" href="run_8h.html#afbf76019c1c088b5b3debd5cc9e9f960">run.h</a>
-</li>
 <li>SCR_TEST_SKIP
 : <a class="el" href="test_8h.html#ae7db7254ee93c4f067e45085a2f0eb7d">test.h</a>
-</li>
-<li>scrCtxCleanupFn
-: <a class="el" href="run_8h.html#a7082a4eda78201e36a72cacf9fe43372">run.h</a>
-</li>
-<li>scrCtxCreateFn
-: <a class="el" href="run_8h.html#ab73c6d249ac70b6eb7f69f6a01ae783b">run.h</a>
-</li>
-<li>scrGroup
-: <a class="el" href="run_8h.html#ac6666827e21523e7bc29f69e17650ef9">run.h</a>
-</li>
-<li>scrGroupAddTest()
-: <a class="el" href="run_8h.html#a99abef0928dfbbd9e32f11c1a3f03308">run.h</a>
 </li>
 <li>scrGroupCreate()
 : <a class="el" href="run_8h.html#a842be1690e36fd4845db6abb8c64c843">run.h</a>
@@ -194,9 +179,6 @@ $(function() {
 </li>
 <li>scrStats
 : <a class="el" href="run_8h.html#a0ab54a6f620fd8de5f8f12228a905406">run.h</a>
-</li>
-<li>scrTestFn
-: <a class="el" href="run_8h.html#afcafbbb6a6f2f1a25b24463d43db05af">run.h</a>
 </li>
 <li>SCRUTINY_VERSION
 : <a class="el" href="scrutiny_8h.html#af8d9e35da7f9738df5ec67f565a96e13">scrutiny.h</a>
