+++ docs/globals_defs.html
@@ -156,9 +156,6 @@ $(function() {
 <li>SCR_RUN_FLAG_FAIL_FAST
 : <a class="el" href="run_8h.html#a4701a8c215ac188f58bead0aaaa2e12d">run.h</a>
 </li>
-<li>SCR_TEST_FLAG_XFAIL
-: <a class="el" href="run_8h.html#afbf76019c1c088b5b3debd5cc9e9f960">run.h</a>
-</li>
 <li>SCR_TEST_SKIP
 : <a class="el" href="test_8h.html#ae7db7254ee93c4f067e45085a2f0eb7d">test.h</a>
 </li>
