+++ docs/globals_func.html
@@ -64,9 +64,6 @@ $(function() {
 
 <div class="contents">
 &#160;<ul>
-<li>scrGroupAddTest()
-: <a class="el" href="run_8h.html#a99abef0928dfbbd9e32f11c1a3f03308">run.h</a>
-</li>
 <li>scrGroupCreate()
 : <a class="el" href="run_8h.html#a842be1690e36fd4845db6abb8c64c843">run.h</a>
 </li>
