+++ docs/globals_type.html
@@ -64,15 +64,6 @@ $(function() {
 
 <div class="contents">
 &#160;<ul>
-<li>scrCtxCleanupFn
-: <a class="el" href="run_8h.html#a7082a4eda78201e36a72cacf9fe43372">run.h</a>
-</li>
-<li>scrCtxCreateFn
-: <a class="el" href="run_8h.html#ab73c6d249ac70b6eb7f69f6a01ae783b">run.h</a>
-</li>
-<li>scrGroup
-: <a class="el" href="run_8h.html#ac6666827e21523e7bc29f69e17650ef9">run.h</a>
-</li>
 <li>scrOptions
 : <a class="el" href="run_8h.html#a86df619ddcdab198dd59ea817de931e6">run.h</a>
 </li>
@@ -81,9 +72,6 @@ $(function() {
 </li>
 <li>scrStats
 : <a class="el" href="run_8h.html#a0ab54a6f620fd8de5f8f12228a905406">run.h</a>
-</li>
-<li>scrTestFn
-: <a class="el" href="run_8h.html#afcafbbb6a6f2f1a25b24463d43db05af">run.h</a>
 </li>
 </ul>
 </div><!-- contents -->
