+++ docs/run_8h__dep__incl.dot
@@ -3,7 +3,7 @@ digraph "include/scrutiny/run.h"
  // LATEX_PDF_SIZE
   edge [fontname="Helvetica",fontsize="10",labelfontname="Helvetica",labelfontsize="10"];
   node [fontname="Helvetica",fontsize="10",shape=record];
-  Node1 [label="include/scrutiny/run.h",height=0.2,width=0.4,color="black", fillcolor="grey75", style="filled", fontcolor="black",tooltip="Defines the runner and group functionality."];
+  Node1 [label="include/scrutiny/run.h",height=0.2,width=0.4,color="black", fillcolor="grey75", style="filled", fontcolor="black",tooltip="Defines the runner functionality."];
   Node1 -> Node2 [dir="back",color="midnightblue",fontsize="10",style="solid",fontname="Helvetica"];
   Node2 [label="include/scrutiny/scrutiny.h",height=0.2,width=0.4,color="black", fillcolor="white", style="filled",URL="$scrutiny_8h.html",tooltip="Master include file."];
 }
