+++ docs/scrutiny_8h__incl.dot
@@ -5,11 +5,13 @@ digraph "include/scrutiny/scrutiny.h"
   node [fontname="Helvetica",fontsize="10",shape=record];
   Node1 [label="include/scrutiny/scrutiny.h",height=0.2,width=0.4,color="black", fillcolor="grey75", style="filled", fontcolor="black",tooltip="Master include file."];
   Node1 -> Node2 [color="midnightblue",fontsize="10",style="solid",fontname="Helvetica"];
-  Node2 [label="run.h",height=0.2,width=0.4,color="black", fillcolor="white", style="filled",URL="$run_8h.html",tooltip="Defines the runner and group functionality."];
-  Node1 -> Node3 [color="midnightblue",fontsize="10",style="solid",fontname="Helvetica"];
-  Node3 [label="test.h",height=0.2,width=0.4,color="black", fillcolor="white", style="filled",URL="$test_8h.html",tooltip="Defines the test macros."];
-  Node3 -> Node4 [color="midnightblue",fontsize="10",style="solid",fontname="Helvetica"];
-  Node4 [label="stdint.h",height=0.2,width=0.4,color="grey75", fillcolor="white", style="filled",tooltip=" "];
-  Node3 -> Node5 [color="midnightblue",fontsize="10",style="solid",fontname="Helvetica"];
-  Node5 [label="stdlib.h",height=0.2,width=0.4,color="grey75", fillcolor="white", style="filled",tooltip=" "];
+  Node2 [label="run.h",height=0.2,width=0.4,color="black", fillcolor="white", style="filled",URL="$run_8h.html",tooltip="Defines the runner functionality."];
+  Node2 -> Node3 [color="midnightblue",fontsize="10",style="solid",fontname="Helvetica"];
+  Node3 [label="groups.h",height=0.2,width=0.4,color="black", fillcolor="white", style="filled",URL="$groups_8h_source.html",tooltip=" "];
+  Node1 -> Node4 [color="midnightblue",fontsize="10",style="solid",fontname="Helvetica"];
+  Node4 [label="test.h",height=0.2,width=0.4,color="black", fillcolor="white", style="filled",URL="$test_8h.html",tooltip="Defines the test macros."];
+  Node4 -> Node5 [color="midnightblue",fontsize="10",style="solid",fontname="Helvetica"];
+  Node5 [label="stdint.h",height=0.2,width=0.4,color="grey75", fillcolor="white", style="filled",tooltip=" "];
+  Node4 -> Node6 [color="midnightblue",fontsize="10",style="solid",fontname="Helvetica"];
+  Node6 [label="stdlib.h",height=0.2,width=0.4,color="grey75", fillcolor="white", style="filled",tooltip=" "];
 }
