+++ docs/scrutiny_8h_source.html
@@ -81,7 +81,7 @@ $(function() {
 <div class="line"><a name="l00015"></a><span class="lineno">   15</span>&#160;<span class="preprocessor">#include &quot;<a class="code" href="test_8h.html">test.h</a>&quot;</span></div>
 <div class="line"><a name="l00016"></a><span class="lineno">   16</span>&#160; </div>
 <div class="line"><a name="l00017"></a><span class="lineno">   17</span>&#160;<span class="preprocessor">#endif  </span><span class="comment">// SCRUTINY_SCRUTINY_H</span></div>
-<div class="ttc" id="arun_8h_html"><div class="ttname"><a href="run_8h.html">run.h</a></div><div class="ttdoc">Defines the runner and group functionality.</div></div>
+<div class="ttc" id="arun_8h_html"><div class="ttname"><a href="run_8h.html">run.h</a></div><div class="ttdoc">Defines the runner functionality.</div></div>
 <div class="ttc" id="atest_8h_html"><div class="ttname"><a href="test_8h.html">test.h</a></div><div class="ttdoc">Defines the test macros.</div></div>
 </div><!-- fragment --></div><!-- contents -->
 <!-- start footer part -->
