+++ docs/search/all_4.js
@@ -30,21 +30,15 @@ var searchData=
   ['scr_5fgroup_5fctx_34',['SCR_GROUP_CTX',['../test_8h.html#a6798a5522104cae2a2588ae48f5fe47f',1,'test.h']]],
   ['scr_5flog_35',['SCR_LOG',['../test_8h.html#ad9fd77add8be18fc2d8737516e209125',1,'test.h']]],
   ['scr_5frun_5fflag_5ffail_5ffast_36',['SCR_RUN_FLAG_FAIL_FAST',['../run_8h.html#a4701a8c215ac188f58bead0aaaa2e12d',1,'run.h']]],
-  ['scr_5ftest_5fflag_5fxfail_37',['SCR_TEST_FLAG_XFAIL',['../run_8h.html#afbf76019c1c088b5b3debd5cc9e9f960',1,'run.h']]],
-  ['scr_5ftest_5fskip_38',['SCR_TEST_SKIP',['../test_8h.html#ae7db7254ee93c4f067e45085a2f0eb7d',1,'test.h']]],
-  ['scrctxcleanupfn_39',['scrCtxCleanupFn',['../run_8h.html#a7082a4eda78201e36a72cacf9fe43372',1,'run.h']]],
-  ['scrctxcreatefn_40',['scrCtxCreateFn',['../run_8h.html#ab73c6d249ac70b6eb7f69f6a01ae783b',1,'run.h']]],
-  ['scrgroup_41',['scrGroup',['../run_8h.html#ac6666827e21523e7bc29f69e17650ef9',1,'run.h']]],
-  ['scrgroupaddtest_42',['scrGroupAddTest',['../run_8h.html#a99abef0928dfbbd9e32f11c1a3f03308',1,'run.h']]],
-  ['scrgroupcreate_43',['scrGroupCreate',['../run_8h.html#a842be1690e36fd4845db6abb8c64c843',1,'run.h']]],
-  ['scroptions_44',['scrOptions',['../run_8h.html#a86df619ddcdab198dd59ea817de931e6',1,'scrOptions():&#160;run.h'],['../structscrOptions.html',1,'scrOptions']]],
-  ['scrrunner_45',['scrRunner',['../run_8h.html#ac3a848eb41861eaa58b64799f782e607',1,'run.h']]],
-  ['scrrunnercreate_46',['scrRunnerCreate',['../run_8h.html#a3a23e3114a02fe1c65e82ef9e4a5514a',1,'run.h']]],
-  ['scrrunnerdestroy_47',['scrRunnerDestroy',['../run_8h.html#a70b2cfa0baa4b8741203eb9c3497098a',1,'run.h']]],
-  ['scrrunnerrun_48',['scrRunnerRun',['../run_8h.html#a449e0477a531f06e7c35ea2fbb2199a1',1,'run.h']]],
-  ['scrstats_49',['scrStats',['../run_8h.html#a0ab54a6f620fd8de5f8f12228a905406',1,'scrStats():&#160;run.h'],['../structscrStats.html',1,'scrStats']]],
-  ['scrtestfn_50',['scrTestFn',['../run_8h.html#afcafbbb6a6f2f1a25b24463d43db05af',1,'run.h']]],
-  ['scrutiny_51',['Scrutiny',['../index.html',1,'']]],
-  ['scrutiny_2eh_52',['scrutiny.h',['../scrutiny_8h.html',1,'']]],
-  ['scrutiny_5fversion_53',['SCRUTINY_VERSION',['../scrutiny_8h.html#af8d9e35da7f9738df5ec67f565a96e13',1,'scrutiny.h']]]
+  ['scr_5ftest_5fskip_37',['SCR_TEST_SKIP',['../test_8h.html#ae7db7254ee93c4f067e45085a2f0eb7d',1,'test.h']]],
+  ['scrgroupcreate_38',['scrGroupCreate',['../run_8h.html#a842be1690e36fd4845db6abb8c64c843',1,'run.h']]],
+  ['scroptions_39',['scrOptions',['../run_8h.html#a86df619ddcdab198dd59ea817de931e6',1,'scrOptions():&#160;run.h'],['../structscrOptions.html',1,'scrOptions']]],
+  ['scrrunner_40',['scrRunner',['../run_8h.html#ac3a848eb41861eaa58b64799f782e607',1,'run.h']]],
+  ['scrrunnercreate_41',['scrRunnerCreate',['../run_8h.html#a3a23e3114a02fe1c65e82ef9e4a5514a',1,'run.h']]],
+  ['scrrunnerdestroy_42',['scrRunnerDestroy',['../run_8h.html#a70b2cfa0baa4b8741203eb9c3497098a',1,'run.h']]],
+  ['scrrunnerrun_43',['scrRunnerRun',['../run_8h.html#a449e0477a531f06e7c35ea2fbb2199a1',1,'run.h']]],
+  ['scrstats_44',['scrStats',['../run_8h.html#a0ab54a6f620fd8de5f8f12228a905406',1,'scrStats():&#160;run.h'],['../structscrStats.html',1,'scrStats']]],
+  ['scrutiny_45',['Scrutiny',['../index.html',1,'']]],
+  ['scrutiny_2eh_46',['scrutiny.h',['../scrutiny_8h.html',1,'']]],
+  ['scrutiny_5fversion_47',['SCRUTINY_VERSION',['../scrutiny_8h.html#af8d9e35da7f9738df5ec67f565a96e13',1,'scrutiny.h']]]
 ];
