+++ docs/search/files_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['scrutiny_2eh_58',['scrutiny.h',['../scrutiny_8h.html',1,'']]]
+  ['scrutiny_2eh_52',['scrutiny.h',['../scrutiny_8h.html',1,'']]]
 ];
