+++ docs/search/functions_0.js
@@ -1,8 +1,7 @@
 var searchData=
 [
-  ['scrgroupaddtest_60',['scrGroupAddTest',['../run_8h.html#a99abef0928dfbbd9e32f11c1a3f03308',1,'run.h']]],
-  ['scrgroupcreate_61',['scrGroupCreate',['../run_8h.html#a842be1690e36fd4845db6abb8c64c843',1,'run.h']]],
-  ['scrrunnercreate_62',['scrRunnerCreate',['../run_8h.html#a3a23e3114a02fe1c65e82ef9e4a5514a',1,'run.h']]],
-  ['scrrunnerdestroy_63',['scrRunnerDestroy',['../run_8h.html#a70b2cfa0baa4b8741203eb9c3497098a',1,'run.h']]],
-  ['scrrunnerrun_64',['scrRunnerRun',['../run_8h.html#a449e0477a531f06e7c35ea2fbb2199a1',1,'run.h']]]
+  ['scrgroupcreate_54',['scrGroupCreate',['../run_8h.html#a842be1690e36fd4845db6abb8c64c843',1,'run.h']]],
+  ['scrrunnercreate_55',['scrRunnerCreate',['../run_8h.html#a3a23e3114a02fe1c65e82ef9e4a5514a',1,'run.h']]],
+  ['scrrunnerdestroy_56',['scrRunnerDestroy',['../run_8h.html#a70b2cfa0baa4b8741203eb9c3497098a',1,'run.h']]],
+  ['scrrunnerrun_57',['scrRunnerRun',['../run_8h.html#a449e0477a531f06e7c35ea2fbb2199a1',1,'run.h']]]
 ];
