+++ docs/search/pages_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['scrutiny_111',['Scrutiny',['../index.html',1,'']]]
+  ['scrutiny_99',['Scrutiny',['../index.html',1,'']]]
 ];
