+++ docs/search/typedefs_0.js
@@ -1,10 +1,6 @@
 var searchData=
 [
-  ['scrctxcleanupfn_71',['scrCtxCleanupFn',['../run_8h.html#a7082a4eda78201e36a72cacf9fe43372',1,'run.h']]],
-  ['scrctxcreatefn_72',['scrCtxCreateFn',['../run_8h.html#ab73c6d249ac70b6eb7f69f6a01ae783b',1,'run.h']]],
-  ['scrgroup_73',['scrGroup',['../run_8h.html#ac6666827e21523e7bc29f69e17650ef9',1,'run.h']]],
-  ['scroptions_74',['scrOptions',['../run_8h.html#a86df619ddcdab198dd59ea817de931e6',1,'run.h']]],
-  ['scrrunner_75',['scrRunner',['../run_8h.html#ac3a848eb41861eaa58b64799f782e607',1,'run.h']]],
-  ['scrstats_76',['scrStats',['../run_8h.html#a0ab54a6f620fd8de5f8f12228a905406',1,'run.h']]],
-  ['scrtestfn_77',['scrTestFn',['../run_8h.html#afcafbbb6a6f2f1a25b24463d43db05af',1,'run.h']]]
+  ['scroptions_64',['scrOptions',['../run_8h.html#a86df619ddcdab198dd59ea817de931e6',1,'run.h']]],
+  ['scrrunner_65',['scrRunner',['../run_8h.html#ac3a848eb41861eaa58b64799f782e607',1,'run.h']]],
+  ['scrstats_66',['scrStats',['../run_8h.html#a0ab54a6f620fd8de5f8f12228a905406',1,'run.h']]]
 ];
