+++ docs/search/variables_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['flags_65',['flags',['../structscrOptions.html#a8467b749302edfa3c9edbba81ad66c44',1,'scrOptions']]]
+  ['flags_58',['flags',['../structscrOptions.html#a8467b749302edfa3c9edbba81ad66c44',1,'scrOptions']]]
 ];
