+++ docs/search/variables_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['global_5fctx_66',['global_ctx',['../structscrOptions.html#a2a2eee9cdea6949ae9b6db842eaf3c09',1,'scrOptions']]]
+  ['global_5fctx_59',['global_ctx',['../structscrOptions.html#a2a2eee9cdea6949ae9b6db842eaf3c09',1,'scrOptions']]]
 ];
