+++ docs/search/variables_2.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['num_5ferrored_67',['num_errored',['../structscrStats.html#a32944f34261641031f5a06f51f786065',1,'scrStats']]],
-  ['num_5ffailed_68',['num_failed',['../structscrStats.html#a2cab04795b417dacc314938b25a6f89b',1,'scrStats']]],
-  ['num_5fpassed_69',['num_passed',['../structscrStats.html#ae1897529c0e662fd86856911a88c6b8b',1,'scrStats']]],
-  ['num_5fskipped_70',['num_skipped',['../structscrStats.html#ad3370711e149678702e565fc920133a0',1,'scrStats']]]
+  ['num_5ferrored_60',['num_errored',['../structscrStats.html#a32944f34261641031f5a06f51f786065',1,'scrStats']]],
+  ['num_5ffailed_61',['num_failed',['../structscrStats.html#a2cab04795b417dacc314938b25a6f89b',1,'scrStats']]],
+  ['num_5fpassed_62',['num_passed',['../structscrStats.html#ae1897529c0e662fd86856911a88c6b8b',1,'scrStats']]],
+  ['num_5fskipped_63',['num_skipped',['../structscrStats.html#ad3370711e149678702e565fc920133a0',1,'scrStats']]]
 ];
